Add handler test
